+++ Nest-Backend/src/students/controller/student.controller.spec.ts
@@ -11,6 +11,7 @@ describe('StudentController', () => {
   const responseMock = {
     status: jest.fn().mockReturnThis(),
     json: jest.fn().mockReturnThis(),
+    send: jest.fn().mockReturnThis(),
   } as unknown as Response;
 
   beforeEach(async () => {
