+++ Nest-Backend/src/students/student.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SocketGateway } from 'src/utils/socket.gateway';
 import { StudentController } from './controller/student.controller';
-import { StudentEntity } from './models/student.entity';
+import { StudentEntity } from './entities/student.entity';
 import { StudentService } from './service/student.service';
 
 @Module({
