--- Nest-Backend/src/user/models/user.interface.ts
@@ -1,6 +1,0 @@
-export interface User {
-  id?: number;
-  email: string;
-  password: string;
-  role: string;
-}
